--- src/components/example-form/styles.module.css
@@ -1,8 +1,0 @@
-.form {
-	margin: 10px;
-  border: 0;
-  border-radius: 3px;
-  box-shadow: 0 3px 5px 2px rgba(0, 0, 0, 0.3);
-  color: black;
-  padding: 30px;
-}
