+++ src/components/index.ts
@@ -1,6 +1,5 @@
 export { BaseButton } from "./base-button/index.tsx";
 export { CustomFormGroup } from "./custom-form-group/index.tsx";
-export { ExampleForm } from "./example-form/example-form.tsx";
 export { PrivateRoute } from "./private-route/index.tsx";
 export { PublicRoute } from "./public-route/index.tsx";
 export { RouterProvider } from "./router-provider/index.tsx";
