+++ src/libs/hooks/index.ts
@@ -1,2 +1,1 @@
-export { useGetUser } from "./use-get-user.ts";
 export { useToggle } from "./use-toggle.ts";
