--- src/libs/types/auth/index.ts
@@ -1,1 +1,0 @@
-export { type AuthTokenResponse } from "../auth/auth-token-response.ts";
