--- src/libs/types/user/user-sign-in-request-dto.ts
@@ -1,6 +1,0 @@
-type UserSignInRequestDto = {
-	email: string;
-	password: string;
-};
-
-export { type UserSignInRequestDto };
