--- src/libs/types/user/user-sign-in-response-dto.ts
@@ -1,9 +1,0 @@
-import { type User } from "./user.ts";
-
-type UserSignInResponseDto = {
-	accessToken: string;
-	refreshToken: string;
-	user: User;
-};
-
-export { type UserSignInResponseDto };
