--- src/libs/types/user/user-sign-up-form-data.ts
@@ -1,8 +1,0 @@
-type UserSignUpFormData = {
-	email: string;
-	name: string;
-	password: string;
-	repeatPassword: string;
-};
-
-export { type UserSignUpFormData };
