--- src/libs/types/user/user-sign-up-request-dto.ts
@@ -1,7 +1,0 @@
-type UserSignUpRequestDto = {
-	email: string;
-	name: string;
-	password: string;
-};
-
-export { type UserSignUpRequestDto };
