--- src/libs/types/user/user-sign-up-response-dto.ts
@@ -1,9 +1,0 @@
-import { type User } from "./user.ts";
-
-type UserSignUpResponseDto = {
-	accessToken: string;
-	refreshToken: string;
-	user: User;
-};
-
-export { type UserSignUpResponseDto };
