--- src/libs/types/user/user.ts
@@ -1,6 +1,0 @@
-type User = {
-	email: string;
-	name: string;
-};
-
-export { type User };
