+++ src/pages/auth/components/index.ts
