+++ src/pages/auth/components/sign-up-link.tsx
