+++ src/pages/auth/components/styles.ts
