--- src/pages/auth/example-styles.module.css
@@ -1,5 +1,0 @@
-.example-selector{
-    width: 100%;
-    display: flex;
-    justify-content: center;
-}
