--- src/pages/auth/libs/constants/index.ts
@@ -1,1 +1,0 @@
-export { exceptionMessage } from "./exception-message.ts";
