--- src/pages/auth/libs/validation/constants/email-regex.ts
@@ -1,3 +1,0 @@
-const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]{2,}$/;
-
-export { emailRegex };
