--- src/pages/auth/libs/validation/constants/index.ts
@@ -1,3 +1,0 @@
-export { emailRegex } from "./email-regex.ts";
-export { userValidationMessage } from "./user-validation-message.ts";
-export { userValidationRules } from "./user-validation-rules.ts";
