--- src/pages/auth/libs/validation/constants/user-validation-rules.ts
@@ -1,6 +1,0 @@
-const userValidationRules = {
-	NAME_MIN_LENGTH: 2,
-	PASSWORD_MIN_LENGTH: 8,
-};
-
-export { userValidationRules };
