--- src/redux/auth/constants/index.ts
@@ -1,1 +1,0 @@
-export { authApiPath } from "./auth-api-path.ts";
