+++ src/redux/hooks.ts
@@ -6,5 +6,5 @@ import {
 
 import type { AppDispatch, RootState } from "./store.ts";
 
-export const useAppDispatch: () => AppDispatch = useDispatch;
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
+export const useAppDispatch: () => AppDispatch = useDispatch;
