--- src/redux/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useAppDispatch } from "./use-app-dispatch.ts";
-export { useAppSelector } from "./use-app-selector.ts";
