--- src/redux/hooks/use-app-dispatch.ts
@@ -1,5 +1,0 @@
-import { useDispatch } from "react-redux";
-
-import type { AppDispatch } from "../store.ts";
-
-export const useAppDispatch: () => AppDispatch = useDispatch;
