--- src/redux/hooks/use-app-selector.ts
@@ -1,5 +1,0 @@
-import { type TypedUseSelectorHook, useSelector } from "react-redux";
-
-import type { RootState } from "../store.ts";
-
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
